+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add tests with ats in ci pipeline.
+- Add helm chart templating test in ci pipeline.
+
 ## [0.3.1] - 2024-07-22
 
 ### Fixed