+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add tests with ats in ci pipeline.
 - Add helm chart templating test in ci pipeline.
+- Add e2e tests. 
 - Add support for Alloy to be used as monitoring agent in-place of Prometheus Agent. This is configurable via the `--monitoring-agent` flag.
 
 ## [0.3.1] - 2024-07-22